+++ Frontend/src/lib/components/index.ts
@@ -4,6 +4,8 @@ export {default as Header} from './layout/Header.svelte';
 export {default as Footer} from './layout/Footer.svelte';
 
 export {default as Card} from './main/Card.svelte';
+export {default as Steps} from './main/Steps.svelte';
+export {default as StepsItem} from './main/StepsItem.svelte';
 export {default as SelectMenu} from './main/SelectMenu.svelte';
 
 export {default as Slogan} from './homepage/Slogan.svelte';