+++ Frontend/src/lib/components/main/Card.svelte
@@ -6,7 +6,7 @@
     upperClass?: string;
     animate?: boolean;
     isMouseEntered?: boolean;
-    children?: Snippet;
+    children: Snippet;
   }
 
   let {color, className, upperClass, animate, isMouseEntered, children}: Props = $props();