+++ Frontend/src/lib/components/main/Steps.svelte
@@ -1,0 +1,19 @@
+<script lang="ts">
+  import {currentStep} from '$store';
+  import type {Snippet} from 'svelte';
+  import {fly} from 'svelte/transition';
+
+  let {children}: {children: Snippet} = $props();
+</script>
+
+{#key $currentStep}
+  <div in:fly={{delay: 500, x: 100, opacity: 0, duration: 500}} out:fly={{x: -100, opacity: 0, duration: 500}}>
+    {@render children?.()}
+  </div>
+{/key}
+
+<style lang="postcss">
+  div {
+    @apply mx-auto flex h-screen w-2/3 justify-center gap-12;
+  }
+</style>