+++ Frontend/src/routes/signup/+page.svelte
@@ -1,19 +1,24 @@
 <script lang="ts">
   import type {Notification, Registration} from '$types';
-  import {enhance} from '$app/forms';
-  import {Card} from '$components';
+  import {Steps, StepsItem, SelectMenu} from '$components';
+  import {currentStep} from '$store';
   import {notify} from '$lib';
 
   const {form}: {form: Notification & Registration} = $props();
 
-  let username = $derived(form?.username);
-  let backup = $derived(form?.backup);
-  let secret = $derived(form?.secret);
-  let step = $derived(form?.step);
-  let qr = $derived(form?.qr);
+  let username = $state() as string;
+  let backup = $state() as string[];
+  let secret = $state() as string;
+  let qr = $state() as string;
 
   $effect(() => {
     if (form?.message) notify(form.message, form.type);
+    if (form?.username) username = form.username;
+    if (form?.backup) backup = form.backup;
+    if (form?.secret) secret = form.secret;
+    if (form?.qr) qr = form.qr;
+
+    currentStep.set(Number(form?.step) || 1);
   });
 </script>
 
@@ -22,44 +27,41 @@
   <meta name="description" content="Sign up and create your account to start using ShadowSelf today" />
 </svelte:head>
 
-<div id="signup">
-  <Card upperClass="w-fit flex self-center">
-    <form method="POST" action="?/init" use:enhance>
-      <h1>Create an account</h1>
-      <div>
-        <label for="username">Username</label>
-        <input type="text" name="username" required id="username" />
+<Steps>
+  <StepsItem index={1} action="init">
+    <h1>Create an account</h1>
+    <div class="flex gap-6">
+      <label for="username">Username</label>
+      <input type="text" name="username" required id="username" />
+    </div>
+    <div class="flex gap-6">
+      <label for="password">Password</label>
+      <input type="password" name="password" required id="password" />
+    </div>
+    <button type="submit">Sign Up</button>
+  </StepsItem>
+  <StepsItem index={2} action="askOTP">
+    <h1>Hello {username}, Would you like to enable 2FA to your account</h1>
+    <button name="add" type="submit">Add</button>
+    <button name="skip" type="submit">Skip</button>
+  </StepsItem>
+  <StepsItem index={3} action="showOTP">
+    {#if qr && secret}
+      <img src={qr} alt="QR Code" width="300" />
+      <div class="flex w-full flex-col gap-4">
+        <p>Scan the QR code with your authenticator app to add 2 factor authentication:</p>
+        <p>alternatively, you can use the code: <b>{secret}</b></p>
+        <p class="text-red-500">make sure to use 'SHA512' as the algorithm</p>
       </div>
-      <div>
-        <label for="password">Password</label>
-        <input type="password" name="password" required id="password" />
-      </div>
-      <button type="submit">Sign Up</button>
-    </form>
-  </Card>
-  <Card upperClass="w-fit flex self-center">
-    <form method="POST" action="?/askOTP" use:enhance>
-      <h1>hello {username}, Add 2FA to your account</h1>
-      {#if qr && secret}
-        <img src={qr} alt="QR Code" width="300" />
-        <div class="flex w-full flex-col gap-4">
-          <p>Scan the QR code with your authenticator app to add 2 factor authentication:</p>
-          <p>alternatively, you can use the code: <b>{secret}</b></p>
-          <p class="text-red-500">make sure to use 'SHA512' as the algorithm</p>
-        </div>
-      {/if}
-      <button name="add" type="submit">Add</button>
-      <button name="skip" type="submit">Skip</button>
-    </form>
-  </Card>
-  <Card upperClass="w-fit flex self-center">
-    <form method="POST" action="?/checkOTP" use:enhance>
-      <h1>Enter the verification code</h1>
-      <input type="number" name="code" class="w-full" required />
-      <button type="submit">Verify</button>
-    </form>
-  </Card>
-  <Card upperClass="w-fit flex self-center">
+    {/if}
+    <button type="submit">Next</button>
+  </StepsItem>
+  <StepsItem index={4} action="checkOTP">
+    <h1>Enter the verification code</h1>
+    <input type="number" name="code" class="w-full" required />
+    <button type="submit">Verify</button>
+  </StepsItem>
+  <StepsItem index={5} action="showBackup">
     <h1>Please Store these Backup Codes Safely</h1>
     <div class="flex flex-col gap-4 rounded-xl bg-slate-900/50 p-4">
       {#if backup}
@@ -70,39 +72,44 @@
       <button>Copy</button>
       <button>Download</button>
     </div>
-  </Card>
-  <Card upperClass="w-fit flex self-center">
-    <form method="POST" action="?/create" use:enhance>
-      <h1>Create account</h1>
-      <button type="submit">Create</button>
-    </form>
-  </Card>
-</div>
+    <button type="submit">Next</button>
+  </StepsItem>
+  <StepsItem index={6} action="askBilling">
+    <h1>Would you like to add a billing method now?</h1>
+    <button name="add" type="submit">Add</button>
+    <button name="skip" type="submit">Skip</button>
+  </StepsItem>
+  <StepsItem index={7} action="addBilling">
+    <h1>Add billing method</h1>
+    <SelectMenu
+      name="Plan"
+      options={[
+        {label: 'Monthly', value: 'monthly'},
+        {label: 'Annually', value: 'annually'},
+        {label: 'Lifetime', value: 'lifetime'},
+      ]} />
+    <button type="submit">Add</button>
+  </StepsItem>
+  <StepsItem index={8} action="create">
+    <h1>Create account</h1>
+    <button type="submit">Create</button>
+  </StepsItem>
+</Steps>
 
 <style lang="postcss">
-  #signup {
-    @apply mx-auto flex h-screen w-2/3 justify-center gap-12;
-  }
-
   h1 {
-    @apply -mt-4 mb-4 self-center text-4xl font-bold text-neutral-300;
+    @apply -mt-4 mb-4 self-center text-center text-4xl font-bold text-neutral-300;
   }
 
   img {
     @apply self-center shadow-lg shadow-white/15;
   }
 
-  form {
-    @apply flex flex-col items-end gap-8 px-12 py-16;
+  input {
+    @apply bg-slate-900/50;
+  }
 
-    & input {
-      @apply bg-slate-900/50;
-    }
-    & div {
-      @apply flex gap-6;
-    }
-    & button {
-      @apply w-full self-end;
-    }
+  button {
+    @apply w-full self-end;
   }
 </style>