+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.08] - 2024-11-24
+
+- Add Step component
+- Add StepItem component
+- Add optional field to check function
+- Add next/back control button to signup
+- Add signup text content
+
 ## [0.2.07] - 2024-11-22
 
 ### Added
 
-- Add Step component
 - Add signup server actions
 - Add control buttons to signup